AWS Cloud Solution Architect Project
The customer had a need to run containers on AWS and to use the same orchestration device for both Cloud and on-premises. It was opted for a seamless transition to cloud-native services as demonstrated visually on the file. It was also needed to host a SQL database and to store that generated base with minimum refactoring