Using SNS
In this model, the client had a business which many of the server workload was intrincately coupled, in other words, if one part of the server failed, all work would be interrupted. On top of that, the monitoring was made in several different manual spreadsheets and there was much operational overhead, with limited staff which was directly impacting on performance. To solve all those issues, we used decoupling functions with Lambda, so that the system would have high resilience to fallovers of any kind. On top of that it was introduced a central monitorizing with cloudwatch and optimizing cost and reducing operational overhead with serverless cloud services.